requirements.txt
Flask==2.3.3
aiohttp==3.9.1
aiogram==3.22.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.4.0
